ai/customer_chatbot/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# LangChain & LangGraph
langgraph>=0.0.40
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-cohere>=0.1.0
cohere>=5.0.0
langchain-community>=0.0.10

# Embeddings
sentence-transformers>=2.2.0
torch>=2.0.0

# Database
psycopg2-binary>=2.9.9
pgvector>=0.2.4
asyncpg>=0.29.0

# Utilities
httpx>=0.25.0
python-multipart>=0.0.6
